summary_table wrapper function to create data.table on target variables using calc_stats

summary_table wrapper function to create data.table on target variables using calc_stats

Usage

summary_table(
  dt,
  target,
  treat,
  target_name = target,
  indent = nbsp(n = 4L),
  .total_dt = NULL,
  pct_dec = 1,
  treat_order = NULL,
  skip_absent = TRUE
)

Arguments

dt

table to perform function on

target

vector of column names desired to obtain information on

treat

string of treatment variable used for splitting / grouping data

target_name

names to be displayed for target columns either NULL, which uses target, of vector of desired names for all target columns

indent

indent to be used for display and formatting purposes

.total_dt

optional table for total counts to be derived

pct_dec

decimal places for percentages

treat_order

customise the column order of output table

skip_absent

Logical, default TRUE. Passed to data.table::setcolorder, if treat_order includes columns not present in dt, TRUE will silently ignore them, FALSE will throw an error.

Value

a data.table containing summary information on target variables specified

Examples

vars<-c('AGE','RACE','ETHNIC','BMRKR1')
var_labels <- c("Age (yr)", "Race", "Ethnicity", "Continous Level Biomarker 1")
DMT01<-summary_table(adsl, target = vars, treat = 'ARM', target_name = var_labels, indent = '  ')
DMT01_pct<-summary_table(adsl, target = vars, treat = 'ARM', indent = '  ', .total_dt = adsl)